requirements.txt
streamlit 
pillow 
numpy 
